# feat(web): use babel-loader instead of ts-loader for web build builder