--- a/packages/react/src/migrations/update-8-5-0/update-workspace-8-5-0.ts
+++ b/packages/react/src/migrations/update-8-5-0/update-workspace-8-5-0.ts
@@ -1,0 +1,24 @@
+import { Rule } from '@angular-devkit/schematics';
+import { updateWorkspaceInTree } from '@nrwl/workspace';
+
+export default function update(): Rule {
+  return updateWorkspaceInTree(config => {
+    const a = [];
+    const b = [];
+    Object.keys(config.schematics).forEach(name => {
+      if (name === '@nrwl/react' && config.schematics[name].application) {
+        a.push(config.schematics[name]);
+      }
+      if (name === '@nrwl/react:application') {
+        b.push(config.schematics[name]);
+      }
+    });
+    a.forEach(x => {
+      delete x.application.babel;
+    });
+    b.forEach(x => {
+      delete x.babel;
+    });
+    return config;
+  });
+}