--- a/packages/web/src/migrations/update-8-5-0/update-builder-8-5-0.ts
+++ b/packages/web/src/migrations/update-8-5-0/update-builder-8-5-0.ts
@@ -1,0 +1,21 @@
+import { Rule } from '@angular-devkit/schematics';
+import { updateWorkspaceInTree } from '@nrwl/workspace';
+
+export default function update(): Rule {
+  return updateWorkspaceInTree(config => {
+    const filteredProjects = [];
+    Object.keys(config.projects).forEach(name => {
+      if (
+        config.projects[name].architect &&
+        config.projects[name].architect.build &&
+        config.projects[name].architect.build.builder === '@nrwl/web:build'
+      ) {
+        filteredProjects.push(config.projects[name]);
+      }
+    });
+    filteredProjects.forEach(p => {
+      delete p.architect.build.options.differentialLoading;
+    });
+    return config;
+  });
+}