--- a/packages/web/src/utils/devserver.config.spec.ts
+++ b/packages/web/src/utils/devserver.config.spec.ts
@@ -19,7 +19,6 @@ describe('getDevServerConfig', () => {
   beforeEach(() => {
     buildInput = {
       main: 'main.ts',
-      differentialLoading: true,
       index: 'index.html',
       budgets: [],
       baseHref: '/',