… docs/angular/api-react/schematics/application.md
… docs/angular/api-web/builders/build.md
… docs/react/api-react/schematics/application.md
… docs/react/api-web/builders/build.md
… docs/web/api-react/schematics/application.md
… docs/web/api-web/builders/build.md
… e2e/react.test.ts
… e2e/web.test.ts
… package.json
… packages/node/package.json
… packages/react/migrations.json
… packages/react/src/migrations/update-8-5-0/update-workspace-8-5-0.spec.ts
… packages/react/src/plugins/babel.ts
… packages/react/src/schematics/application/application.spec.ts
… packages/react/src/schematics/application/schema.json
… packages/react/src/utils/versions.ts
… packages/web/migrations.json
… packages/web/package.json
… packages/web/src/builders/build/build.impl.spec.ts
… packages/web/src/builders/build/schema.json
… packages/web/src/migrations/update-8-5-0/update-builder-8-5-0.spec.ts
… packages/web/src/utils/config.spec.ts
… packages/web/src/utils/web.config.spec.ts
… yarn.lock