## Changes
- removes the `differentialLoading` build option
- differential loading is always enabled for prod builds

## Benchmarks

Created a react app with 100 component libs and 100 redux slices.

```
for X in {1..100}
do
  nx g lib "page-$X" -a $*
  nx g slice "slice-$X" -p $*
done
```

With differential loading on, we've increased the build time by 5 seconds on average.

BEFORE (without ESM):
```
Benchmark #1: nx build demo --prod
  Time (mean ± σ):     13.834 s ±  1.731 s    [User: 11.817 s, System: 1.352 s]
  Range (min … max):   11.947 s … 16.015 s    10 runs
```
AFTER (with ESM):
```
Benchmark #1: nx build demo --prod
  Time (mean ± σ):     18.711 s ±  1.310 s    [User: 12.172 s, System: 1.394 s]
  Range (min … max):   17.232 s … 20.770 s    10 runs
```